pipeline/requirements.txt
fastapi
uvicorn[standard]
httpx
websockets
Pillow
